+++ edb/edgeql-parser/edgeql-parser-python/src/parser.rs
@@ -5,10 +5,7 @@ use cpython::{
     PythonObject, PythonObjectWithCheckedDowncast, ToPyObject,
 };
 
-use edgeql_parser::keywords;
 use edgeql_parser::parser;
-use edgeql_parser::tokenizer;
-use indexmap::IndexMap;
 
 use crate::errors::{parser_error_into_tuple, ParserResult};
 use crate::pynormalize::value_to_py_object;
@@ -73,12 +70,7 @@ pub fn downcast_tokens<'a>(py: Python, token_list: PyObject) -> PyResult<Vec<par
     for token in tokens.iter(py) {
         let token = OpaqueToken::downcast_from(py, token)?.inner(py);
 
-        buf.push(parser::Terminal {
-            kind: token.kind,
-            text: token.text,
-            value: token.value,
-            span: token.span,
-        });
+        buf.push(parser::Terminal::from_token(token));
     }
     Ok(buf)
 }
@@ -120,116 +112,17 @@ fn load_spec(py: Python, parser_name: &str) -> PyResult<&'static (parser::Spec, 
 
         let spec_json =
             PyString::downcast_from(py, res.get_item(py, 0)).expect("json to be a string");
+        let spec_json = spec_json.to_string(py).unwrap();
+        std::fs::write(parser_name, spec_json.as_bytes()).unwrap();
+
         let productions = res.get_item(py, 1);
 
-        let spec = spec_from_json(&spec_json.to_string(py).unwrap()).unwrap();
+        let spec = parser::Spec::from_json(&spec_json).unwrap();
 
         parser_specs.insert(parser_name.to_string(), (spec, productions));
     }
 
     Ok(unsafe { PARSER_SPECS.as_ref().unwrap().get(parser_name).unwrap() })
-}
-
-pub fn spec_from_json(j_spec: &str) -> Result<parser::Spec, String> {
-    use serde::Deserialize;
-
-    #[derive(Debug, Deserialize)]
-    pub struct SpecJson {
-        pub actions: Vec<Vec<(String, parser::Action)>>,
-        pub goto: Vec<Vec<(String, usize)>>,
-        pub start: String,
-        pub inlines: Vec<(usize, u8)>,
-    }
-
-    let v = serde_json::from_str::<SpecJson>(j_spec).map_err(|e| e.to_string())?;
-
-    let actions = v
-        .actions
-        .into_iter()
-        .map(|x| x.into_iter().map(|(k, a)| (get_token_kind(&k), a)))
-        .map(IndexMap::from_iter)
-        .collect();
-    let goto = v.goto.into_iter().map(IndexMap::from_iter).collect();
-    let inlines = IndexMap::from_iter(v.inlines);
-    Ok(parser::Spec {
-        actions,
-        goto,
-        start: v.start,
-        inlines,
-    })
-}
-
-fn get_token_kind(token_name: &str) -> tokenizer::Kind {
-    use tokenizer::Kind::*;
-
-    match token_name {
-        "+" => Add,
-        "&" => Ampersand,
-        "@" => At,
-        ".<" => BackwardLink,
-        "}" => CloseBrace,
-        "]" => CloseBracket,
-        ")" => CloseParen,
-        "??" => Coalesce,
-        ":" => Colon,
-        "," => Comma,
-        "++" => Concat,
-        "/" => Div,
-        "." => Dot,
-        "**" => DoubleSplat,
-        "=" => Eq,
-        "//" => FloorDiv,
-        "%" => Modulo,
-        "*" => Mul,
-        "::" => Namespace,
-        "{" => OpenBrace,
-        "[" => OpenBracket,
-        "(" => OpenParen,
-        "|" => Pipe,
-        "^" => Pow,
-        ";" => Semicolon,
-        "-" => Sub,
-
-        "?!=" => DistinctFrom,
-        ">=" => GreaterEq,
-        "<=" => LessEq,
-        "?=" => NotDistinctFrom,
-        "!=" => NotEq,
-        "<" => Less,
-        ">" => Greater,
-
-        "IDENT" => Ident,
-        "EOF" => EOF,
-        "<$>" => EOI,
-        "<e>" => Epsilon,
-
-        "BCONST" => BinStr,
-        "FCONST" => FloatConst,
-        "ICONST" => IntConst,
-        "NFCONST" => DecimalConst,
-        "NICONST" => BigIntConst,
-        "SCONST" => Str,
-
-        "+=" => AddAssign,
-        "->" => Arrow,
-        ":=" => Assign,
-        "-=" => SubAssign,
-
-        "ARGUMENT" => Argument,
-        "SUBSTITUTION" => Substitution,
-
-        _ => {
-            let mut token_name = token_name.to_lowercase();
-
-            if let Some(rem) = token_name.strip_prefix("dunder") {
-                token_name = format!("__{rem}__");
-            }
-
-            let kw = keywords::lookup_all(&token_name)
-                .unwrap_or_else(|| panic!("unknown keyword {token_name}"));
-            Keyword(kw)
-        }
-    }
 }
 
 fn to_py_cst<'a>(cst: parser::CSTNode, py: Python) -> PyResult<CSTNode> {
